HW10: rename /internal/domain to /internal/interfaces
