+++ hw12_13_14_15_calendar/internal/logger_uber/logger_uber.go
@@ -2,7 +2,7 @@ package logger_uber
 
 import (
 	"fmt"
-	"github.com/artembert/golang-pro-otus-hw/hw12_13_14_15_calendar/internal/domain/loglevel"
+	"github.com/artembert/golang-pro-otus-hw/hw12_13_14_15_calendar/internal/interfaces/loglevel"
 	"go.uber.org/zap"
 )
