+++ hw12_13_14_15_calendar/internal/storage/sql/storage.go
@@ -3,7 +3,7 @@ package sqlstorage
 import (
 	"context"
 	"github.com/artembert/golang-pro-otus-hw/hw12_13_14_15_calendar/domain"
-	"github.com/artembert/golang-pro-otus-hw/hw12_13_14_15_calendar/internal/domain/storage"
+	"github.com/artembert/golang-pro-otus-hw/hw12_13_14_15_calendar/internal/interfaces/storage"
 	"github.com/jackc/pgx/v4/pgxpool"
 )
